+++ src/audionode.rs
@@ -913,7 +913,7 @@ where
     }
 
     fn process(&mut self, size: usize, input: &BufferRef, output: &mut BufferMut) {
-        let mut buffer = BufferArray::<X::Outputs>::new();
+        let mut buffer = BufferArray::<X::Outputs>::uninitialized();
         self.x.process(
             size,
             &input.subset(0, self.x.inputs()),
@@ -1361,7 +1361,7 @@ where
     }
 
     fn process(&mut self, size: usize, input: &BufferRef, output: &mut BufferMut) {
-        let mut buffer = BufferArray::<X::Outputs>::new();
+        let mut buffer = BufferArray::<X::Outputs>::uninitialized();
         self.x.process(size, input, &mut buffer.buffer_mut());
         self.y.process(size, &buffer.buffer_ref(), output);
     }
@@ -1714,7 +1714,7 @@ where
     }
 
     fn process(&mut self, size: usize, input: &BufferRef, output: &mut BufferMut) {
-        let mut buffer = BufferArray::<X::Outputs>::new();
+        let mut buffer = BufferArray::<X::Outputs>::uninitialized();
         self.x.process(size, input, output);
         self.y.process(size, input, &mut buffer.buffer_mut());
         for channel in 0..self.outputs() {
@@ -1804,7 +1804,7 @@ impl<X: AudioNode> AudioNode for Thru<X> {
         if X::Inputs::USIZE < X::Outputs::USIZE {
             // An intermediate buffer is only used in this "degenerate" case where
             // we are not passing through inputs - we are cutting out some of them.
-            let mut buffer = BufferArray::<X::Outputs>::new();
+            let mut buffer = BufferArray::<X::Outputs>::uninitialized();
             self.x.process(size, input, &mut buffer.buffer_mut());
             for channel in 0..X::Inputs::USIZE {
                 for i in 0..simd_items(size) {
@@ -1906,7 +1906,7 @@ where
     }
 
     fn process(&mut self, size: usize, input: &BufferRef, output: &mut BufferMut) {
-        let mut buffer = BufferArray::<X::Outputs>::new();
+        let mut buffer = BufferArray::<X::Outputs>::uninitialized();
         self.x[0].process(size, input, output);
         for i in 1..N::USIZE {
             self.x[i].process(size, input, &mut buffer.buffer_mut());
@@ -2174,7 +2174,7 @@ where
         output
     }
     fn process(&mut self, size: usize, input: &BufferRef, output: &mut BufferMut) {
-        let mut buffer = BufferArray::<X::Outputs>::new();
+        let mut buffer = BufferArray::<X::Outputs>::uninitialized();
         self.x[0].process(size, &input.subset(0, X::Inputs::USIZE), output);
         let mut in_channel = X::Inputs::USIZE;
         for i in 1..N::USIZE {
@@ -2417,7 +2417,7 @@ where
     }
 
     fn process(&mut self, size: usize, input: &BufferRef, output: &mut BufferMut) {
-        let mut buffer = BufferArray::<X::Outputs>::new();
+        let mut buffer = BufferArray::<X::Outputs>::uninitialized();
         if N::USIZE & 1 > 0 {
             self.x[0].process(size, input, output);
         } else {